requirements.txt
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
DateTime==5.2
h11==0.14.0
idna==3.4
jsonpath-ng==1.6.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
outcome==1.2.0
ply==3.11
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.30.2
requests==2.31.0
robotframework==6.1.1
robotframework-jsonlibrary==0.5
robotframework-pythonlibcore==4.2.0
robotframework-requests==0.9.5
robotframework-seleniumlibrary==6.1.2
rpds-py==0.10.4
selenium==4.13.0
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.11.1
urllib3==2.0.6
wsproto==1.2.0
zope.interface==6.1
